+++ apps/elf-eval/src/bin/real_world_live_adapter/runtime_support/commands.rs
@@ -1,0 +1,82 @@
+use std::{
+	fs::{self, OpenOptions},
+	io::Write as _,
+	path::Path,
+	process::{Command, Stdio},
+};
+
+use color_eyre::{Result, eyre};
+
+use crate::QmdArgs;
+
+pub(crate) fn run_qmd_command(
+	label: &str,
+	args: &QmdArgs,
+	home_dir: &Path,
+	qmd_args: &[&str],
+	log_path: &Path,
+) -> Result<String> {
+	let mut command = Command::new("npx");
+
+	command
+		.current_dir(&args.qmd_dir)
+		.env("HOME", home_dir)
+		.env("XDG_CACHE_HOME", "/root/.cache")
+		.env("QMD_FORCE_CPU", "1")
+		.arg("tsx")
+		.arg("src/cli/qmd.ts");
+
+	for arg in qmd_args {
+		command.arg(arg);
+	}
+
+	run_logged_command(label, &mut command, log_path)
+}
+
+pub(crate) fn run_logged_shell(
+	label: &str,
+	cwd: &Path,
+	script: &str,
+	log_path: &Path,
+) -> Result<()> {
+	let mut command = Command::new("bash");
+
+	command.current_dir(cwd).arg("-lc").arg(script);
+
+	run_logged_command(label, &mut command, log_path).map(|_| ())
+}
+
+pub(crate) fn run_logged_command(
+	label: &str,
+	command: &mut Command,
+	log_path: &Path,
+) -> Result<String> {
+	if let Some(parent) = log_path.parent() {
+		fs::create_dir_all(parent)?;
+	}
+
+	let command_debug = format!("{command:?}");
+	let output = command.stdout(Stdio::piped()).stderr(Stdio::piped()).output()?;
+	let stdout = String::from_utf8_lossy(&output.stdout).to_string();
+	let stderr = String::from_utf8_lossy(&output.stderr).to_string();
+	let mut log = OpenOptions::new().create(true).append(true).open(log_path)?;
+
+	writeln!(log, "## {label}")?;
+	writeln!(log, "$ {command_debug}")?;
+
+	if !stdout.trim().is_empty() {
+		writeln!(log, "\nstdout:\n{stdout}")?;
+	}
+	if !stderr.trim().is_empty() {
+		writeln!(log, "\nstderr:\n{stderr}")?;
+	}
+	if !output.status.success() {
+		return Err(eyre::eyre!(
+			"{label} failed with status {}. Inspect {}.",
+			output.status,
+			log_path.display()
+		));
+	}
+
+	Ok(stdout)
+}
