+++ apps/elf-eval/src/bin/real_world_live_adapter/runtime_support/config.rs
@@ -1,0 +1,35 @@
+use std::sync::Arc;
+
+use color_eyre::Result;
+
+use crate::{BaselineRuntime, DeterministicEmbedding, NoopExtractor, TokenOverlapRerank};
+use elf_config::Config;
+use elf_service::Providers;
+
+pub(crate) fn runtime_config(runtime: &BaselineRuntime) -> Result<Config> {
+	let mut cfg = elf_config::load(&runtime.config_path)?;
+
+	cfg.storage.postgres.dsn = runtime.dsn.clone();
+	cfg.storage.postgres.pool_max_conns = 12;
+	cfg.storage.qdrant.url = runtime.qdrant_url.clone();
+	cfg.storage.qdrant.collection = runtime.collection.clone();
+	cfg.storage.qdrant.docs_collection = runtime.docs_collection.clone();
+	cfg.providers.embedding.provider_id = "local".to_string();
+	cfg.providers.embedding.model = "local-hash".to_string();
+	cfg.providers.embedding.dimensions = cfg.storage.qdrant.vector_dim;
+	cfg.providers.rerank.provider_id = "local".to_string();
+	cfg.providers.rerank.model = "local-token-overlap".to_string();
+	cfg.providers.llm_extractor.provider_id = "disabled".to_string();
+	cfg.providers.llm_extractor.model = "disabled".to_string();
+	cfg.context = None;
+
+	Ok(cfg)
+}
+
+pub(crate) fn deterministic_providers(vector_dim: u32) -> Providers {
+	Providers::new(
+		Arc::new(DeterministicEmbedding { vector_dim }),
+		Arc::new(TokenOverlapRerank),
+		Arc::new(NoopExtractor),
+	)
+}
