+++ apps/elf-eval/src/bin/real_world_live_adapter/runtime_support/embedding.rs
@@ -1,0 +1,48 @@
+use std::collections::BTreeSet;
+
+pub(crate) fn embed_text(text: &str, vector_dim: u32) -> Vec<f32> {
+	let dim = vector_dim as usize;
+	let mut vector = vec![0.0_f32; dim];
+
+	if dim == 0 {
+		return vector;
+	}
+
+	let normalized = normalize_ascii_alnum_lowercase(text);
+
+	for term in normalized.split_whitespace() {
+		if term.len() < 2 {
+			continue;
+		}
+
+		let hash = blake3::hash(term.as_bytes());
+		let bytes = hash.as_bytes();
+		let idx = (u32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]) as usize) % dim;
+
+		vector[idx] += 1.0;
+	}
+
+	let norm = vector.iter().map(|value| value * value).sum::<f32>().sqrt();
+
+	if norm > 0.0 {
+		for value in &mut vector {
+			*value /= norm;
+		}
+	}
+
+	vector
+}
+
+pub(crate) fn terms(text: &str) -> BTreeSet<String> {
+	normalize_ascii_alnum_lowercase(text)
+		.split_whitespace()
+		.filter(|term| term.len() >= 2)
+		.map(ToString::to_string)
+		.collect()
+}
+
+pub(crate) fn normalize_ascii_alnum_lowercase(text: &str) -> String {
+	text.chars()
+		.map(|ch| if ch.is_ascii_alphanumeric() { ch.to_ascii_lowercase() } else { ' ' })
+		.collect()
+}
